refactor: chat 컴포넌트 분리
- redux slice 추가 생성: userListSlice, selectedUserSlice
