+++ src/redux/index.ts
@@ -1,11 +1,17 @@
 import { configureStore } from "@reduxjs/toolkit";
 import { createLogger } from "redux-logger";
 import userReducers from "./slice/userSlice";
+import userListReducers from "./slice/userListSlice";
+import selectedUserReducers from "./slice/selectedUserSlice";
 
 const logger = createLogger();
 
 const store = configureStore({
-  reducer: { user: userReducers },
+  reducer: {
+    user: userReducers,
+    userList: userListReducers,
+    selectedUser: selectedUserReducers,
+  },
   middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(logger),
   devTools: true,
 });
